Change some return type.

+++ src/list.h
@@ -179,7 +179,7 @@ typedef struct list_head
  * \param head the list
  * \return size of the list
  */
-static inline int list_size(struct list_head* head)
+static inline size_t list_size(struct list_head* head)
 {
   struct list_head* lp = NULL;
   size_t size = 0;
